Refactoring shooter lookup table, in prep to remove sorted vector
